RONDE 177: het bijknippen blijft nu ook echt staan
De trim schreef altijd al weg. Wat eromheen zat, deed dat niet.

1. De previewlink veranderde nooit
   /api/admin/archive/media/<id> is voor en na het bijknippen exact dezelfde
   string, en die route vraagt de browser hem een uur te bewaren
   (private, max-age=3600; 300s op de S3-redirect). Dus: rij opnieuw opgehaald,
   video uit de cache van de browser gespeeld — de oude, ongeknipte beelden.
   Alles werkte en er was niets van te zien.

   De link draagt nu een versietoken dat verandert wanneer storagePut nieuwe
   bytes wegschrijft, en op geen enkel ander moment. Eén definitie in
   shared/archiveMediaVersion.ts, gedeeld door de archiefgrid, beide
   streamroutes en het editormanifest.

2. durationSec was een INT-kolom
   De trim meet het bestand dat hij zojuist maakte en schrijft 8,53s terug;
   MySQL rondde dat af naar 9. Het getal in de rij was dus nooit het getal dat
   de gebruiker te zien kreeg, en een trim die minder dan een halve seconde
   afhaalde schreef hetzelfde getal terug — een rij die er onaangeraakt uitzag.
   Migratie 0048 verbreedt de kolom naar FLOAT; media_asset_cache.durationSec
   was daar al FLOAT voor.

3. De oordelen over de oude beelden bleven staan
   annotationJson, annotationVersion, hasBakedEditText, previewCheckedAt en
   previewIssue zijn conclusies uit de frames, en de frames zijn net veranderd.
   Ze worden in dezelfde update leeggemaakt; NULL betekent bij alle vijf al
   "nog niet bekeken", dus de bestaande annotator en de bestaande controles
   kijken vanzelf opnieuw. Geen nieuwe checker.

   editorialScore blijft expres staan: leeg wordt in de montagefilter als 50
   gelezen, dus leegmaken zou elke net bijgeknipte clip omlaag duwen. De
   annotator schrijft hem bij zijn volgende ronde toch over.

29 nieuwe tests, waarvan de trim er met echte ffmpeg, echte storagePut en
echte ffprobe doorheen gaat. Zes mutaties gedraaid (id-only link, constant
token, ontbrekende dependency, INT-kolom terug, annotatie niet leeggemaakt,
editormanifest zonder rij) — alle zes gepakt.

5225 tests groen, tsc schoon, eslint 0 fouten, build groen.

+++ drizzle/schema.ts
@@ -179,7 +179,17 @@ export const mediaArchiveAssets = mysqlTable(
   topics: json("topics").$type<string[]>(),
   width: int("width"),
   height: int("height"),
-  durationSec: int("durationSec"),
+  /**
+   * RONDE 177 — FLOAT, not INT.
+   *
+   * The trim writes the probed duration of the file it just produced, which is fractional
+   * (8.53s). An INT column rounds that on the way in, so the number in the row disagreed with the
+   * number the trim returned to the operator — and a trim that shortened a clip by less than half
+   * a second wrote back the SAME integer, leaving a row that looked untouched.
+   *
+   * media_asset_cache.durationSec was already FLOAT for exactly this reason.
+   */
+  durationSec: float("durationSec"),
   sortOrder: int("sortOrder").default(0).notNull(),
   isActive: int("isActive").default(1).notNull(),
   /** Cached overlay-filter verdict: null = not yet checked, 0 = clean, 1 = baked edit text detected. */
